[builtin] Implement builtin sleep
Also fix time_::sleep() in C++ to take a double.  It's implemented with
nanosleep().

TODO:

- document it
- add 'builtin cat', and remove __cat
- implement the rest of 'invoke'
  - --show should be a table
  - implement --builtin --extern --proc-like

+++ cpp/stdlib.h
@@ -170,7 +170,7 @@ time_t time();
 // factor out our own API with better types.
 time_t localtime(time_t ts);
 BigStr* strftime(BigStr* s, time_t ts);
-void sleep(int seconds);
+void sleep(double seconds);
 
 }  // namespace time_
